Cleaned up type/function names, added documentation, and unified namespaces for all Rfws library
