+++ Source/Components/RfwsSerialization/Properties/AssemblyInfo.cs
